requirements.txt
opencv-python~=4.6.0.66
matplotlib~=3.5.2
pillow~=9.2.0